+++ R/count_na.R
@@ -56,7 +56,7 @@ count_na <- function(x, ...) {
     for (i in colnames(.dat)) {
       # print freq
       dummy <- count_na_helper(.dat[[i]])
-      cat(sprintf("# %s\n\n", get_label(.dat[[i]], def.value = i)))
+      cat(sprintf("# %s\n\n", sjlabelled::get_label(.dat[[i]], def.value = i)))
       print(dummy)
       cat("\n\n")
       # save data frame for return value
@@ -69,7 +69,7 @@ count_na <- function(x, ...) {
     dummy <- count_na_helper(.dat)
 
     # check if we have variable label and print, if yes
-    vl <- get_label(.dat)
+    vl <- sjlabelled::get_label(.dat)
     if (!is.null(vl)) cat(sprintf("# %s\n\n", vl))
 
     # print count table
@@ -82,6 +82,7 @@ count_na <- function(x, ...) {
 }
 
 
+#' @importFrom sjlabelled get_na
 count_na_helper <- function(x) {
   # check if x has any tagged NA values
   if (sum(haven::is_tagged_na(x)) < 1) {
@@ -90,8 +91,8 @@ count_na_helper <- function(x) {
   }
 
   # get NA as tagged NA
-  nav <- haven::na_tag(get_na(x, as.tag = F))
-  nav.labels <- names(get_na(x, as.tag = T))
+  nav <- haven::na_tag(sjlabelled::get_na(x, as.tag = F))
+  nav.labels <- names(sjlabelled::get_na(x, as.tag = T))
   # get values from x, including different NA tags
   values <- haven::na_tag(x)
   # only keep missing values